Unset BUNDLE_GEMFILE when using tmux
If tmux is started in a ruby project using rvm and a Gemfile, and then
a second tmux session is started, the second tmux session will have the
environment variables set from the first session.

One of the env variables is BUNDLE_GEMFILE which confuses rvm when trying
to find gems from another project.

Unsetting this env variable in .bashrc seems to fix this problem.

https://github.com/tmuxinator/tmuxinator/issues/161